requirements.txt
exa-py==1.14.16
python-dotenv==1.1.1 